% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfbeta_plot.R
\name{arg_check_dfbeta_plot}
\alias{arg_check_dfbeta_plot}
\title{Check arguments of dfbeta_plot}
\usage{
arg_check_dfbeta_plot(
  model,
  id_n,
  regressors,
  add_reference,
  text_arglist,
  abline_arglist,
  extendrange_f
)
}
\arguments{
\item{model}{A fitted model object from the
\code{\link[stats]{lm}} function.}

\item{id_n}{The number of points to identify with labels.
The default is \code{3}.}

\item{add_reference}{A logical value indicating whether a
reference line should be added. The default is
\code{TRUE}.}

\item{text_arglist}{Additional arguments passed to the
\code{\link[graphics]{text}} function, which is used to
display the points that are identified.}

\item{abline_arglist}{A named list specifying additional
arguments passed to the \code{\link[graphics]{abline}}
function for the horizontal reference line added to the
plot.}

\item{extendrange_f}{Positive number(s) specifying the
fraction by which the range of the residuals should be
extended using the \code{\link[grDevices]{extendrange}}
function. If longer than one, \code{f[1]} is used on
the left and \code{f[2]} on the right.}
}
\description{
Check arguments of dfbeta_plot
}
\keyword{internal}
