% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffits_test.R
\name{arg_check_dffits_test}
\alias{arg_check_dffits_test}
\title{Check arguments of \code{dffits_test}}
\usage{
arg_check_dffits_test(model, n)
}
\arguments{
\item{model}{A fitted model object from the
\code{\link[stats]{lm}} function.}

\item{n}{The number of outliers to return. The default is
all influential observations.}
}
\description{
Check arguments of \code{dffits_test}
}
\keyword{internal}
