% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_plot.R
\name{arg_check_influence_plot}
\alias{arg_check_influence_plot}
\title{Check arguments of dfbetas_plot}
\usage{
arg_check_influence_plot(
  model,
  id_n,
  add_reference,
  alpha,
  size,
  text_arglist,
  abline_arglist,
  extendrange_f
)
}
\arguments{
\item{model}{A fitted model object from the
\code{\link[stats]{lm}} function.}

\item{id_n}{The number of points to identify with labels
with respect to largest absolute criterion. The default
is \code{3} labels.}

\item{add_reference}{A logical value indicating whether a
reference line should be added. The default is
\code{TRUE}.}

\item{alpha}{The default quantile used for the horizontal
reference lines. The default is 0.05. See Details.}

\item{size}{A numeric vector of length 2 that provides
guidelines for the size of the points.}

\item{text_arglist}{Additional arguments passed to the
\code{\link[graphics]{text}} function, which is used to
display the points that are identified.}

\item{abline_arglist}{A named list specifying additional
arguments passed to the \code{\link[graphics]{abline}}
function for the horizontal reference line added to the
plot.}

\item{extendrange_f}{Positive number(s) specifying the
fraction by which the range of the residuals should be
extended using the \code{\link[grDevices]{extendrange}}
function. If longer than one, \code{f[1]} is used on
the left and \code{f[2]} on the right.}
}
\description{
Check arguments of dfbetas_plot
}
\keyword{internal}
