% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leverage_test.R
\name{arg_check_leverage_test}
\alias{arg_check_leverage_test}
\title{Check arguments of \code{leverage_test}}
\usage{
arg_check_leverage_test(model, n, ttype, threshold)
}
\arguments{
\item{model}{A fitted model object from the
\code{\link[stats]{lm}} function.}

\item{n}{The number of leverage points to return. The
default is all leverage points.}

\item{ttype}{Threshold type. The default is
\code{"half"}. The other options are \code{"2mean"} and
\code{"custom"}. See Details.}

\item{threshold}{A number between 0 and 1. Any
observation with a leverage value above this number is
declared a leverage point. This is automatically
determined unless \code{ttype = "custom"}.}
}
\value{
A vector of statistics
}
\description{
Check arguments of \code{leverage_test}
}
