% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_test.R
\name{arg_check_outlier_test}
\alias{arg_check_outlier_test}
\title{Check arguments of \code{outlier_test}}
\usage{
arg_check_outlier_test(model, n, alpha)
}
\arguments{
\item{model}{A fitted model object from the
\code{\link[stats]{lm}} function.}

\item{n}{The number of outliers to return. The default is
all outliers.}

\item{alpha}{The Bonferroni-adjusted threshold at which
an outlier is identified. The default is \code{0.05}.}
}
\description{
Check arguments of \code{outlier_test}
}
\keyword{internal}
