% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_plot_helpers.R
\name{auto_mfrow}
\alias{auto_mfrow}
\title{Automatically determine `mfrow`}
\usage{
auto_mfrow(n)
}
\arguments{
\item{n}{The number of plots.}
}
\value{
A `numeric` vector of length 2.
}
\description{
Automatically determine a reasonable choice for
for `mfrow` in the \code{\link[graphics]{par}}
function based on the number of plots `n`.
}
\keyword{auto_mfrow(4)}
\keyword{internal}
