% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_compare.R
\name{combine_coefs}
\alias{combine_coefs}
\title{Combine coefficients}
\usage{
combine_coefs(i, a, b, digits = NULL)
}
\arguments{
\item{i}{Index of coefficients}

\item{a}{Coefficients and standard errors from Model 1}

\item{b}{Coefficients and standard errors from Model 2}
}
\value{
A character matrix
}
\description{
Combine coefficients and standard errors from two
fitted models
}
\keyword{internal}
