% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_compare.R
\name{compare_coef}
\alias{compare_coef}
\title{Compare coefficients of 2 models}
\usage{
compare_coef(model1, model2, digits = 3)
}
\arguments{
\item{model1}{A fitted model object from the
\code{\link[stats]{lm}} function.}

\item{model2}{A fitted model object from the
\code{\link[stats]{lm}} function.}

\item{inheritParms}{}
}
\value{
A data frame
}
\description{
Compare the coefficients to two fitted models. The models
must have the same coefficients.
}
\examples{
# fit model
lmod1 <- lm(murder ~ hs_grad + urban + poverty + single,
           data = crime2009)
#fit without DC
lmod2 <- lm(murder ~ hs_grad + urban + poverty + single,
            data = crime2009[-9, ])
#compare coefficients of models
compare_compare(lmod1, lmod2)
}
