% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooks_plot.R
\name{cooks_plot}
\alias{cooks_plot}
\title{Index plot of Cook's distances for \code{lm} object}
\usage{
cooks_plot(
  model,
  id_n = 3,
  add_reference = TRUE,
  ...,
  text_arglist = list(),
  abline_arglist = list(),
  extendrange_f = 0.08
)
}
\arguments{
\item{model}{A fitted model object from the
\code{\link[stats]{lm}} function.}

\item{id_n}{The number of points to identify with labels.
The default is \code{3}.}

\item{add_reference}{A logical value indicating whether a
reference line should be added. See details.}

\item{...}{Additional arguments passed to the
\code{\link[graphics]{plot}} function.}

\item{text_arglist}{Additional arguments passed to the
\code{\link[graphics]{text}} function, which is used to
display the points that are identified.}

\item{abline_arglist}{A named list specifying additional
arguments passed to the \code{\link[graphics]{abline}}
function for the horizontal reference line added to the
plot.}

\item{extendrange_f}{Positive number(s) specifying the
fraction by which the range of the residuals should be
extended using the \code{\link[grDevices]{extendrange}}
function. If longer than one, \code{f[1]} is used on
the left and \code{f[2]} on the right.}
}
\description{
\code{cooks_plot} plots the Cook's distances from
the \code{\link[stats]{cooks.distance}} function of a fitted
\code{lm} object.
}
\details{
By default, a reference line is plotted at the \code{0.5}
quantile of a \eqn{F_{p, n-p}} distribution
where \eqn{p = }\code{length(stats::coef(model))} and
\eqn{n - p = }\code{stats::df.residual(model)}.

The vertical position of the reference line can be customized
by setting the \code{h} argument of \code{abline_arglist}.
}
\examples{
lmod <- lm(price ~ sqft_living, data = home_sales)
cooks_plot(lmod, id_n = 1)
}
\seealso{
\code{\link[graphics]{plot}},
  \code{\link[graphics]{text}},
  \code{\link[graphics]{abline}},
  \code{\link[stats]{cooks.distance}}
}
\author{
Joshua French
}
