% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwaine.R
\docType{data}
\name{dwaine}
\alias{dwaine}
\title{Dwaine Studios data}
\format{
A data frame with 21 observations and 3 variables:
\describe{
 \item{\code{targetpop}}{The number of persons aged 16 or younger in thousands of persons.}
 \item{\code{dpi}}{Per capita disposable personal income in thousands of dollars.}
 \item{\code{sales}}{Sales in thousands of dollars}
}
}
\usage{
data(dwaine)
}
\description{
Data from the Dwaine Studios data in Applied
  Linear Statistical Models, 5th edition, p. 237. From
  the book:

  Dwaine Studios, Inc., operates portrait studios in 21
  cities of medium size. These studios specialize in
  portraits of children. The company is considering an
  expansion into other cities of medium size and wishes
  to investigates where sales (\code{sales}) in a
  community can be predicted from the number of persons
  aged 16 or younger in the community (\code{targetpop})
  and the per capita disposable personal income in the
  community (\code{dpi}). Data on these variables for the
  most recent year for the 21 cities in which Dwaine
  Studios is now operating are included in the data set.
}
\references{
Kutner, M. H., Nachtsheim, C. J., Neter, J., & Li, W. (2004). Applied Linear Statistical Models, 5th edition. New York: McGraw-Hill/Irwin.
}
\author{
Joshua P. French
}
\keyword{data}
