% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_residuals.R
\name{get_residuals}
\alias{get_residuals}
\title{Extract residuals from a model}
\usage{
get_residuals(
  x,
  type = c("ordinary", "standardized", "studentized", "jackknife", "loo", "deleted",
    "internally studentized", "externally studentized")
)
}
\arguments{
\item{x}{An `lm` object}

\item{type}{The desired residual type. The options are
`"ordinary"`, `"standardized"`, `"studentized"`,
`"jackknife"`, `"loo"`, `"deleted"`, `"internally
studentized"`, and `"externally studentized"`.}
}
\value{
A vector of residals.
}
\description{
Extract residuals from a model
}
\examples{
lmod <- lm(Girth ~ Height, data = trees)
# ordinary residuals
rord <- get_residuals(lmod)
all.equal(rord, residuals(lmod))
# standardized residuals
rstand <- get_residuals(lmod, "standardized")
all.equal(rstand, rstandard(lmod))
# studentized residuals
rstud <- get_residuals(lmod, "studentized")
all.equal(rstud, rstudent(lmod))
# loo residuals
rl <- get_residuals(lmod, "loo")
all.equal(rl, rloo(lmod))
}
