% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_plot_lm_helpers.R
\name{index_plot_dfbetas_elements}
\alias{index_plot_dfbetas_elements}
\title{Get elements for dfbetas index plot}
\usage{
index_plot_dfbetas_elements(model, returned_stats)
}
\arguments{
\item{model}{A fitted model object from the
\code{\link[stats]{lm}} function.}

\item{returned_stats}{A vector of dfbetas statistics}
}
\value{
A list with elements \code{x}, \code{y},
and \code{z}.
}
\description{
Get elements for dfbetas index plot
}
\keyword{internal}
