% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_plot_lm_helpers.R
\name{index_plot_lm_elements}
\alias{index_plot_lm_elements}
\title{Create plot elements for index_plot_lm}
\usage{
index_plot_lm_elements(model, stat)
}
\arguments{
\item{model}{A fitted model object from the
\code{\link[stats]{lm}} function.}

\item{stat}{A function that can be applied to an \code{lm}
object and returns a vector of observations for each
observation used to fit the model.}
}
\value{
A list with elements \code{x}, \code{y},
and \code{z}.
}
\description{
Create plot elements for index_plot_lm
}
\keyword{internal}
