% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_plot_lm_helpers.R
\name{index_plot_raw}
\alias{index_plot_raw}
\title{Index plot helper function}
\usage{
index_plot_raw(
  x,
  y,
  idd,
  labels,
  add_reference,
  arglist,
  text_arglist,
  abline_arglist,
  extendrange_f
)
}
\arguments{
\item{x}{x-values to plot}

\item{y}{y-values to plot}

\item{idd}{Identified observations}

\item{labels}{The labels to use for the identified points.}

\item{add_reference}{Logical value}

\item{arglist}{Named list for plot}

\item{text_arglist}{Named list for text}

\item{abline_arglist}{Named list for abline}

\item{extendrange_f}{Positive number(s) specifying the
fraction by which the range of the residuals should be
extended using the \code{\link[grDevices]{extendrange}}
function. If longer than one, \code{f[1]} is used on
the left and \code{f[2]} on the right.}
}
\description{
Index plot helper function
}
