% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_plot.R
\name{residual_plot}
\alias{residual_plot}
\title{Plot residuals of a fitted model}
\usage{
residual_plot(model, ...)
}
\arguments{
\item{model}{A fitted model.}

\item{...}{Currently unimplemented.}
}
\description{
\code{residual_plot} plots the residuals of a fitted model.
}
\examples{
lmod <- lm(Girth ~ Height, data = trees)
residual_plot(lmod)
}
\author{
Joshua French
}
