% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rloo.R
\name{rloo}
\alias{rloo}
\alias{rdeleted}
\alias{rjackknife}
\title{Compute leave-one-out residuals}
\usage{
rloo(model, ...)

rdeleted(model, ...)

rjackknife(model, ...)
}
\arguments{
\item{model}{a fitted model object from the \code{\link[stats]{lm}} function.}

\item{...}{Currently unimplemented}
}
\description{
\code{rloo} computes the leave-one-out residuals of
\code{model}. \cr\cr
\code{rjacknife} and \code{rdeleted} are
aliases for \code{rloo}.
}
\examples{
lmod <- lm(Girth ~ Height, data = trees)
rloo(lmod)
}
\seealso{
\code{\link[api2lm]{rloo.lm}}
}
\author{
Joshua French
}
