% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_plot_helpers.R
\name{rplot_raw}
\alias{rplot_raw}
\title{Helper function for residuals_plot.lm}
\usage{
rplot_raw(
  x,
  y,
  idd,
  labels,
  xlab,
  smooth,
  add_reference,
  add_smooth,
  arglist,
  text_arglist,
  abline_arglist,
  smooth_arglist,
  lines_arglist,
  extendrange_f
)
}
\arguments{
\item{x}{x-values to plot}

\item{y}{y-values to plot}

\item{idd}{Identified observations}

\item{labels}{The labels to use for the identified points.}

\item{xlab}{The x-axis label}

\item{smooth}{The function for smoothing. Needs a `formula` argument.}

\item{add_reference}{Logical value}

\item{add_smooth}{Logical value}

\item{arglist}{Named list for plot}

\item{text_arglist}{Named list for text}

\item{abline_arglist}{Named list for abline}

\item{smooth_arglist}{Named list for smooth}

\item{lines_arglist}{Named list for lines of smooth}

\item{extendrange_f}{Positive number(s) specifying the
fraction by which the range of the residuals should be
extended using the \code{\link[grDevices]{extendrange}}
function. If longer than one, \code{f[1]} is used on
the left and \code{f[2]} on the right.}
}
\description{
Helper function for residuals_plot.lm
}
