% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_plot_helpers.R
\name{set_abline_arglist}
\alias{set_abline_arglist}
\title{Set default values of abline_arglist}
\usage{
set_abline_arglist(abline_arglist)
}
\arguments{
\item{abline_arglist}{A named list.}
}
\value{
A named list.
}
\description{
Set the default values of abline_arglist if not
specified by the user.
}
\keyword{internal}
