% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_plot_helpers.R
\name{set_lines_arglist}
\alias{set_lines_arglist}
\title{Set default values of lines_arglist}
\usage{
set_lines_arglist(lines_arglist)
}
\arguments{
\item{lines_arglist}{A named list.}
}
\value{
A named list.
}
\description{
Set the default values of lines_arglist if not
specified by the user.
}
\keyword{internal}
