% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_plot_helpers.R
\name{set_text_arglist}
\alias{set_text_arglist}
\title{Set the default values of text_arglist if not
specified by the user.}
\usage{
set_text_arglist(text_arglist, x, y, labels, idd)
}
\arguments{
\item{text_arglist}{A named list.}

\item{x}{The x-values of the labels.}

\item{y}{The y-values of the labels.}

\item{labels}{The labels to plot.}
}
\value{
A named list.
}
\description{
Set the default values of text_arglist if not
specified by the user.
}
\keyword{internal}
