% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_plot.R
\name{sl_plot}
\alias{sl_plot}
\title{Spread-level plot of a fitted model}
\usage{
sl_plot(model, ...)
}
\arguments{
\item{model}{A fitted model.}

\item{...}{Currently unimplemented.}
}
\description{
\code{sl_plot} creates a spread-level plot
 for a fitted model.
}
\examples{
lmod <- lm(Girth ~ Height, data = trees)
sl_plot(lmod)
}
\author{
Joshua French
}
