% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_stats-summary.R
\name{summary.cor_stats}
\alias{summary.cor_stats}
\alias{print.cor_stats_summary}
\title{Summarize \code{cor_stats} object}
\usage{
\method{summary}{cor_stats}(object, ...)

\method{print}{cor_stats_summary}(x, ..., digits = 2, threshold = 0.7)
}
\arguments{
\item{object}{A \code{cor_stats} object produced by the
\code{\link[api2lm]{cor_stats}} function.}

\item{...}{Not currently implemented}

\item{x}{A `data.frame` or `lm` object.}

\item{digits}{The number of decimal places to print. The
default is \code{2}.}

\item{threshold}{A threshold determining which
correlation statistics to print. Statistics above the
threshold are printed. The default is \code{0.7}.
Otherwise, a "\code{.}" is displayed in place of the
number.}
}
\description{
Summarize an object of class \code{cor_stats} produced
by the \code{\link[api2lm]{cor_stats}} function. Pairs
of variables are ordered in descending order by their
sample correlation.
}
\examples{
cstats <- cor_stats(home_sales)
cstats_sum <- summary(cstats)
cstats_sum
print(cstats_sum, digits = 3, threshold = 0.5)
}
\author{
Joshua French
}
