% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vifstats.R
\name{vif_stats}
\alias{vif_stats}
\title{Compute variance inflation factors}
\usage{
vif_stats(object, ...)
}
\arguments{
\item{object}{A \code{data.frame} or \code{lm} object.}

\item{...}{Not currently implemented.}
}
\value{
A \code{numeric} vector
}
\description{
\code{vif_stats} computes the variance inflation factors
of an object. If \code{object} is a \code{data.frame},
then the variance inflation factors are computed for the
\code{numeric} variables. If \code{object} is an
\code{lm} object, then variance inflation factors for the
matrix of regressor variables not including the
intercept.
}
\examples{
# compute vif stats for a data frame
vif_stats(crime2009)
lmod <- lm(violent ~ hs_grad + poverty + white + single + urban,
           data = crime2009)
# compute vif stats for a fitted model
vif_stats(lmod)
}
