% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_adjust-print.R
\name{print.predict_adjust}
\alias{print.predict_adjust}
\title{Print \code{predict_adjust} object}
\usage{
\method{print}{predict_adjust}(x, ...)
}
\arguments{
\item{x}{An \code{predict_adjust} object produced by the
\code{\link[api2lm]{predict_adjust}} function.}

\item{...}{Additional arguments to the
\code{\link[base]{print.default}} function, such as
\code{digits}.}
}
\description{
Print an object of class \code{predict_adjust} produced
by the \code{\link[api2lm]{predict_adjust}} function.
}
\examples{
fit <- lm(100/mpg ~ disp + hp + wt + am, data = mtcars)
(cia <- predict_adjust(fit))
print(cia, digits = 3)
}
\author{
Joshua French
}
